+++ cpp02/ex02/includes/Fixed.hpp
@@ -6,13 +6,14 @@
 /*   By: mfeldman <mfeldman@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/18 23:24:27 by mfeldman          #+#    #+#             */
-/*   Updated: 2024/02/05 13:08:05 by mfeldman         ###   ########.fr       */
+/*   Updated: 2024/02/05 13:39:28 by mfeldman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #pragma once
 #include <iostream>
 #include <cmath>
+#include <stdint.h>
 
 class Fixed
 {
@@ -52,8 +53,8 @@ class Fixed
 		float	to_float() const;
     	
 	private:
-		int					_raw_bits;
-		static int const	_nb_fractional_bits;
+		int						_raw_bits;
+		static uint8_t const	_nb_fractional_bits;
 };
 
 std::ostream&	operator<<(std::ostream &os, Fixed const &i);
